Extract issue mutation dispatch handlers (Fixes #27)
